Enable GLB Healthcheck to announce VIP/Bind status into BGP
This pull request adds a new feature to GLB healthcheck. It allows it to locally announce the status of a VIP/bind into BGP using a local instance of GoBGP. It integrates with a locally deployed GoBGP daemon to announce the VIP/Bind into BGP when at least one backend is marked healthy. This allows you to prevent a statically announced route via BIRD to black hole traffic for a GLB instance that is unable to talk to its backends. In the event that all backends are marked either inactive or marked as unhealthy GLB healthcheck will withdraw the route. The feature is configurable (see included documentation) and defaults to off. We have been running this in production for a while and it simplified our integration with BGP.
Merge Golang fixes